refactor(gateway): extract NoopSkillsService into services/skills.rs (file-size limit)
