+++ crates/gateway/src/services/skills.rs
@@ -1,0 +1,1408 @@
+use {
+    async_trait::async_trait,
+    serde_json::Value,
+    std::{
+        collections::HashSet,
+        path::{Path, PathBuf},
+        sync::Arc,
+    },
+};
+
+use super::*;
+
+// ── Skills (Noop — complex impl that depends on gateway-specific crates) ────
+
+pub struct NoopSkillsService;
+
+#[async_trait]
+impl SkillsService for NoopSkillsService {
+    async fn status(&self) -> ServiceResult {
+        Ok(serde_json::json!({ "installed": [] }))
+    }
+
+    async fn bins(&self) -> ServiceResult {
+        Ok(serde_json::json!([]))
+    }
+
+    async fn install(&self, params: Value) -> ServiceResult {
+        let source = params
+            .get("source")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'source' parameter (owner/repo format)".to_string())?;
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let skills = moltis_skills::install::install_skill(source, &install_dir)
+            .await
+            .map_err(ServiceError::message)?;
+        let installed: Vec<_> = skills
+            .iter()
+            .map(|m| {
+                serde_json::json!({
+                    "name": m.name,
+                    "description": m.description,
+                    "path": m.path.to_string_lossy(),
+                })
+            })
+            .collect();
+        security_audit(
+            "skills.install",
+            serde_json::json!({
+                "source": source,
+                "installed_count": installed.len(),
+            }),
+        );
+        Ok(serde_json::json!({ "installed": installed }))
+    }
+
+    async fn update(&self, _p: Value) -> ServiceResult {
+        Err("skills not available".into())
+    }
+
+    async fn list(&self) -> ServiceResult {
+        use moltis_skills::{
+            discover::{FsSkillDiscoverer, SkillDiscoverer},
+            requirements::check_requirements,
+        };
+        let fs_discoverer = FsSkillDiscoverer::new(FsSkillDiscoverer::default_paths());
+
+        #[cfg(feature = "bundled-skills")]
+        let skills = {
+            let bundled = Arc::new(moltis_skills::bundled::BundledSkillStore::new());
+            let composite = moltis_skills::discover::CompositeSkillDiscoverer::new(
+                Box::new(fs_discoverer),
+                bundled,
+            );
+            composite.discover().await.map_err(ServiceError::message)?
+        };
+        #[cfg(not(feature = "bundled-skills"))]
+        let skills = fs_discoverer
+            .discover()
+            .await
+            .map_err(ServiceError::message)?;
+        let items: Vec<_> = skills
+            .iter()
+            .map(|s| {
+                let elig = check_requirements(s);
+                let protected = matches!(
+                    s.source,
+                    Some(moltis_skills::types::SkillSource::Personal)
+                        | Some(moltis_skills::types::SkillSource::Project)
+                ) && is_protected_discovered_skill(&s.name);
+                serde_json::json!({
+                    "name": s.name,
+                    "description": s.description,
+                    "category": s.category,
+                    "license": s.license,
+                    "allowed_tools": s.allowed_tools,
+                    "path": s.path.to_string_lossy(),
+                    "source": s.source,
+                    "protected": protected,
+                    "eligible": elig.eligible,
+                    "missing_bins": elig.missing_bins,
+                    "install_options": elig.install_options,
+                })
+            })
+            .collect();
+        Ok(serde_json::json!(items))
+    }
+
+    async fn remove(&self, params: Value) -> ServiceResult {
+        let source = params
+            .get("source")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'source' parameter".to_string())?;
+
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        moltis_skills::install::remove_repo(source, &install_dir)
+            .await
+            .map_err(ServiceError::message)?;
+
+        security_audit("skills.remove", serde_json::json!({ "source": source }));
+
+        Ok(serde_json::json!({ "removed": source }))
+    }
+
+    async fn repos_list(&self) -> ServiceResult {
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let manifest_path = moltis_skills::manifest::ManifestStore::default_path()
+            .map_err(ServiceError::message)?;
+        let store = moltis_skills::manifest::ManifestStore::new(manifest_path);
+        let mut manifest = store.load().map_err(ServiceError::message)?;
+        let (drift_changed, drifted_sources) =
+            detect_and_mark_repo_drift(&mut manifest, &install_dir);
+        if drift_changed {
+            store.save(&manifest).map_err(ServiceError::message)?;
+        }
+
+        // Filter out ClawHub individual skills — they show in the Skills tab, not Repositories.
+        let repos: Vec<_> = manifest
+            .repos
+            .iter()
+            .filter(|repo| !moltis_skills::clawhub::is_clawhub_source(&repo.source))
+            .map(|repo| {
+                let enabled = repo.skills.iter().filter(|s| s.enabled).count();
+                let trusted = repo.skills.iter().filter(|s| s.trusted).count();
+                // Re-detect format for repos that predate the formats module
+                let format = if repo.format == moltis_skills::formats::PluginFormat::Skill {
+                    let repo_dir = install_dir.join(&repo.repo_name);
+                    moltis_skills::formats::detect_format(&repo_dir)
+                } else {
+                    repo.format
+                };
+                serde_json::json!({
+                    "source": repo.source,
+                    "repo_name": repo.repo_name,
+                    "installed_at_ms": repo.installed_at_ms,
+                    "commit_sha": repo.commit_sha,
+                    "quarantined": repo.quarantined,
+                    "quarantine_reason": repo.quarantine_reason,
+                    "provenance": repo.provenance,
+                    "drifted": drifted_sources.contains(&repo.source),
+                    "format": format,
+                    "skill_count": repo.skills.len(),
+                    "enabled_count": enabled,
+                    "trusted_count": trusted,
+                })
+            })
+            .collect();
+
+        let mut repos = repos;
+        if let Ok(entries) = std::fs::read_dir(&install_dir) {
+            for entry in entries.flatten() {
+                let path = entry.path();
+                if !path.is_dir() {
+                    continue;
+                }
+                let repo_name = entry.file_name().to_string_lossy().to_string();
+                if manifest.repos.iter().any(|r| r.repo_name == repo_name) {
+                    continue;
+                }
+                let format = moltis_skills::formats::detect_format(&path);
+                repos.push(serde_json::json!({
+                    "source": format!("orphan:{repo_name}"),
+                    "repo_name": repo_name,
+                    "installed_at_ms": 0,
+                    "commit_sha": null,
+                    "drifted": false,
+                    "orphaned": true,
+                    "format": format,
+                    "skill_count": 0,
+                    "enabled_count": 0,
+                }));
+            }
+        }
+
+        Ok(serde_json::json!(repos))
+    }
+
+    async fn repos_list_full(&self) -> ServiceResult {
+        use moltis_skills::requirements::check_requirements;
+
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let manifest_path = moltis_skills::manifest::ManifestStore::default_path()
+            .map_err(ServiceError::message)?;
+        let store = moltis_skills::manifest::ManifestStore::new(manifest_path);
+        let mut manifest = store.load().map_err(ServiceError::message)?;
+        let (drift_changed, drifted_sources) =
+            detect_and_mark_repo_drift(&mut manifest, &install_dir);
+        if drift_changed {
+            store.save(&manifest).map_err(ServiceError::message)?;
+        }
+
+        let repos: Vec<_> = manifest
+            .repos
+            .iter()
+            .filter(|repo| !moltis_skills::clawhub::is_clawhub_source(&repo.source))
+            .map(|repo| {
+                let repo_dir = install_dir.join(&repo.repo_name);
+                // Re-detect format for repos that predate the formats module
+                let format = if repo.format == moltis_skills::formats::PluginFormat::Skill {
+                    moltis_skills::formats::detect_format(&repo_dir)
+                } else {
+                    repo.format
+                };
+
+                // For non-SKILL.md formats, scan with adapter to get enriched metadata.
+                let adapter_entries = match format {
+                    moltis_skills::formats::PluginFormat::Skill => None,
+                    _ => moltis_skills::formats::scan_with_adapter(&repo_dir, format)
+                        .and_then(|r| r.ok()),
+                };
+
+                let skills: Vec<_> = repo
+                    .skills
+                    .iter()
+                    .map(|s| {
+                        // If we have adapter entries, match by name for enriched data.
+                        if let Some(ref entries) = adapter_entries {
+                            let entry = entries.iter().find(|e| e.metadata.name == s.name);
+                            serde_json::json!({
+                                "name": s.name,
+                                "description": entry.map(|e| e.metadata.description.as_str()).unwrap_or(""),
+                                "display_name": entry.and_then(|e| e.display_name.as_deref()),
+                                "relative_path": s.relative_path,
+                                "trusted": s.trusted,
+                                "enabled": s.enabled,
+                                "drifted": drifted_sources.contains(&repo.source),
+                                "eligible": true,
+                                "missing_bins": [],
+                            })
+                        } else {
+                            // SKILL.md format: parse from disk.
+                            let skill_dir = install_dir.join(&s.relative_path);
+                            let skill_md = skill_dir.join("SKILL.md");
+                            let meta_json = moltis_skills::parse::read_meta_json(&skill_dir);
+                            let (description, display_name, elig) =
+                                if let Ok(content) = std::fs::read_to_string(&skill_md) {
+                                    if let Ok(meta) = moltis_skills::parse::parse_metadata(
+                                        &content, &skill_dir,
+                                    ) {
+                                        let e = check_requirements(&meta);
+                                        let desc = if meta.description.is_empty() {
+                                            meta_json
+                                                .as_ref()
+                                                .and_then(|m| m.display_name.clone())
+                                                .unwrap_or_default()
+                                        } else {
+                                            meta.description
+                                        };
+                                        let dn = meta_json
+                                            .as_ref()
+                                            .and_then(|m| m.display_name.clone());
+                                        (desc, dn, Some(e))
+                                    } else {
+                                        let dn = meta_json
+                                            .as_ref()
+                                            .and_then(|m| m.display_name.clone());
+                                        (dn.clone().unwrap_or_default(), dn, None)
+                                    }
+                                } else {
+                                    let dn =
+                                        meta_json.as_ref().and_then(|m| m.display_name.clone());
+                                    (dn.clone().unwrap_or_default(), dn, None)
+                                };
+                            serde_json::json!({
+                                "name": s.name,
+                                "description": description,
+                                "display_name": display_name,
+                                "relative_path": s.relative_path,
+                                "trusted": s.trusted,
+                                "enabled": s.enabled,
+                                "drifted": drifted_sources.contains(&repo.source),
+                                "eligible": elig.as_ref().map(|e| e.eligible).unwrap_or(true),
+                                "missing_bins": elig.as_ref().map(|e| e.missing_bins.clone()).unwrap_or_default(),
+                            })
+                        }
+                    })
+                    .collect();
+
+                serde_json::json!({
+                    "source": repo.source,
+                    "repo_name": repo.repo_name,
+                    "installed_at_ms": repo.installed_at_ms,
+                    "commit_sha": repo.commit_sha,
+                    "quarantined": repo.quarantined,
+                    "quarantine_reason": repo.quarantine_reason,
+                    "provenance": repo.provenance,
+                    "drifted": drifted_sources.contains(&repo.source),
+                    "format": format,
+                    "skills": skills,
+                })
+            })
+            .collect();
+
+        let mut repos = repos;
+        if let Ok(entries) = std::fs::read_dir(&install_dir) {
+            for entry in entries.flatten() {
+                let path = entry.path();
+                if !path.is_dir() {
+                    continue;
+                }
+                let repo_name = entry.file_name().to_string_lossy().to_string();
+                if manifest.repos.iter().any(|r| r.repo_name == repo_name) {
+                    continue;
+                }
+                let format = moltis_skills::formats::detect_format(&path);
+                repos.push(serde_json::json!({
+                    "source": format!("orphan:{repo_name}"),
+                    "repo_name": repo_name,
+                    "installed_at_ms": 0,
+                    "commit_sha": null,
+                    "drifted": false,
+                    "orphaned": true,
+                    "format": format,
+                    "skills": [],
+                }));
+            }
+        }
+
+        Ok(serde_json::json!(repos))
+    }
+
+    async fn repos_remove(&self, params: Value) -> ServiceResult {
+        let source = params
+            .get("source")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'source' parameter".to_string())?;
+
+        if let Some(repo_name) = source.strip_prefix("orphan:") {
+            let install_dir =
+                moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+            let dir = install_dir.join(repo_name);
+            if dir.exists() {
+                std::fs::remove_dir_all(&dir).map_err(ServiceError::message)?;
+            }
+            security_audit(
+                "skills.orphan.remove",
+                serde_json::json!({ "source": source, "repo_name": repo_name }),
+            );
+            return Ok(serde_json::json!({ "removed": source }));
+        }
+
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        moltis_skills::install::remove_repo(source, &install_dir)
+            .await
+            .map_err(ServiceError::message)?;
+
+        security_audit(
+            "skills.repos.remove",
+            serde_json::json!({ "source": source }),
+        );
+
+        Ok(serde_json::json!({ "removed": source }))
+    }
+
+    async fn repos_export(&self, params: Value) -> ServiceResult {
+        let source = params
+            .get("source")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'source' parameter".to_string())?;
+        let output_path = params
+            .get("path")
+            .and_then(|v| v.as_str())
+            .map(PathBuf::from);
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let exported = moltis_skills::portability::export_repo_bundle(
+            source,
+            &install_dir,
+            output_path.as_deref(),
+        )
+        .await
+        .map_err(ServiceError::message)?;
+
+        security_audit(
+            "skills.repos.export",
+            serde_json::json!({
+                "source": source,
+                "path": exported.bundle_path,
+            }),
+        );
+
+        Ok(serde_json::json!({
+            "source": exported.repo.source,
+            "repo_name": exported.repo.repo_name,
+            "path": exported.bundle_path,
+        }))
+    }
+
+    async fn repos_import(&self, params: Value) -> ServiceResult {
+        let bundle_path = params
+            .get("path")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'path' parameter".to_string())?;
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let imported =
+            moltis_skills::portability::import_repo_bundle(Path::new(bundle_path), &install_dir)
+                .await
+                .map_err(ServiceError::message)?;
+
+        security_audit(
+            "skills.repos.import",
+            serde_json::json!({
+                "source": imported.source,
+                "repo_name": imported.repo_name,
+                "path": imported.bundle_path,
+                "skill_count": imported.skills.len(),
+            }),
+        );
+
+        Ok(serde_json::json!({
+            "source": imported.source,
+            "repo_name": imported.repo_name,
+            "format": imported.format,
+            "path": imported.bundle_path,
+            "quarantined": true,
+            "skill_count": imported.skills.len(),
+            "skills": imported.skills.iter().map(|skill| serde_json::json!({
+                "name": skill.name,
+                "description": skill.description,
+                "path": skill.path.to_string_lossy(),
+            })).collect::<Vec<_>>(),
+        }))
+    }
+
+    async fn repos_unquarantine(&self, params: Value) -> ServiceResult {
+        let source = params
+            .get("source")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'source' parameter".to_string())?;
+
+        let manifest_path = moltis_skills::manifest::ManifestStore::default_path()
+            .map_err(ServiceError::message)?;
+        let store = moltis_skills::manifest::ManifestStore::new(manifest_path);
+        let mut manifest = store.load().map_err(ServiceError::message)?;
+        let repo = manifest
+            .find_repo_mut(source)
+            .ok_or_else(|| format!("repo '{source}' not found"))?;
+        repo.quarantined = false;
+        repo.quarantine_reason = None;
+        store.save(&manifest).map_err(ServiceError::message)?;
+
+        security_audit(
+            "skills.repos.unquarantine",
+            serde_json::json!({ "source": source }),
+        );
+
+        Ok(serde_json::json!({ "source": source, "quarantined": false }))
+    }
+
+    async fn emergency_disable(&self) -> ServiceResult {
+        let manifest_path = moltis_skills::manifest::ManifestStore::default_path()
+            .map_err(ServiceError::message)?;
+        let store = moltis_skills::manifest::ManifestStore::new(manifest_path);
+        let mut manifest = store.load().map_err(ServiceError::message)?;
+
+        let mut disabled = 0_u64;
+        for repo in &mut manifest.repos {
+            for skill in &mut repo.skills {
+                if skill.enabled {
+                    disabled += 1;
+                }
+                skill.enabled = false;
+            }
+        }
+        store.save(&manifest).map_err(ServiceError::message)?;
+
+        security_audit(
+            "skills.emergency_disable",
+            serde_json::json!({ "disabled": disabled }),
+        );
+
+        Ok(serde_json::json!({ "disabled": disabled }))
+    }
+
+    async fn skill_enable(&self, params: Value) -> ServiceResult {
+        toggle_skill(&params, true)
+    }
+
+    async fn skill_disable(&self, params: Value) -> ServiceResult {
+        let source = params.get("source").and_then(|v| v.as_str()).unwrap_or("");
+
+        // Personal/project skills live as files — delete the directory to disable.
+        if source == "personal" || source == "project" {
+            return delete_discovered_skill(source, &params);
+        }
+
+        toggle_skill(&params, false)
+    }
+
+    async fn skill_trust(&self, params: Value) -> ServiceResult {
+        set_skill_trusted(&params, true)
+    }
+
+    /// List bundled skill categories with skill counts and enabled state.
+    async fn bundled_categories(&self) -> ServiceResult {
+        #[cfg(feature = "bundled-skills")]
+        {
+            let store = moltis_skills::bundled::BundledSkillStore::new();
+            let skills = store.discover();
+            let config = moltis_config::discover_and_load();
+            let disabled = &config.skills.disabled_bundled_categories;
+
+            let mut cats: std::collections::BTreeMap<String, u32> =
+                std::collections::BTreeMap::new();
+            for s in &skills {
+                if let Some(cat) = &s.category {
+                    *cats.entry(cat.clone()).or_insert(0) += 1;
+                }
+            }
+
+            let categories: Vec<Value> = cats
+                .into_iter()
+                .map(|(name, count)| {
+                    let enabled = !disabled.iter().any(|d| d == &name);
+                    serde_json::json!({ "name": name, "count": count, "enabled": enabled })
+                })
+                .collect();
+
+            Ok(serde_json::json!({ "categories": categories, "total_skills": skills.len() }))
+        }
+        #[cfg(not(feature = "bundled-skills"))]
+        {
+            Ok(serde_json::json!({ "categories": [], "total_skills": 0 }))
+        }
+    }
+
+    /// Toggle a bundled skill category on or off.
+    async fn bundled_toggle_category(&self, params: Value) -> ServiceResult {
+        let category = params
+            .get("category")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'category' parameter".to_string())?;
+        let enabled = params
+            .get("enabled")
+            .and_then(|v| v.as_bool())
+            .ok_or_else(|| "missing 'enabled' parameter".to_string())?;
+
+        let category = category.to_string();
+        let cat_clone = category.clone();
+
+        if let Err(e) = moltis_config::update_config(|cfg| {
+            if enabled {
+                cfg.skills
+                    .disabled_bundled_categories
+                    .retain(|c| c != &cat_clone);
+            } else if !cfg
+                .skills
+                .disabled_bundled_categories
+                .iter()
+                .any(|c| c == &cat_clone)
+            {
+                cfg.skills
+                    .disabled_bundled_categories
+                    .push(cat_clone.clone());
+            }
+        }) {
+            return Err(format!("failed to save config: {e}").into());
+        }
+
+        Ok(serde_json::json!({ "category": category, "enabled": enabled }))
+    }
+
+    async fn skill_detail(&self, params: Value) -> ServiceResult {
+        use moltis_skills::requirements::check_requirements;
+
+        let source = params
+            .get("source")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'source' parameter".to_string())?;
+        let skill_name = params
+            .get("skill")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'skill' parameter".to_string())?;
+
+        // Personal/project skills: look up directly by name in discovered paths.
+        if source == "personal" || source == "project" {
+            return skill_detail_discovered(source, skill_name);
+        }
+
+        // Bundled skills: read from the embedded store.
+        #[cfg(feature = "bundled-skills")]
+        if source == "bundled" {
+            return skill_detail_bundled(skill_name);
+        }
+
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let manifest_path = moltis_skills::manifest::ManifestStore::default_path()
+            .map_err(ServiceError::message)?;
+        let store = moltis_skills::manifest::ManifestStore::new(manifest_path);
+        let mut manifest = store.load().map_err(ServiceError::message)?;
+        let (drift_changed, drifted_sources) =
+            detect_and_mark_repo_drift(&mut manifest, &install_dir);
+        if drift_changed {
+            store.save(&manifest).map_err(ServiceError::message)?;
+        }
+
+        let repo = manifest
+            .repos
+            .iter()
+            .find(|r| r.source == source)
+            .ok_or_else(|| format!("repo '{source}' not found"))?;
+        let skill_state = repo
+            .skills
+            .iter()
+            .find(|s| s.name == skill_name)
+            .ok_or_else(|| format!("skill '{skill_name}' not found in repo '{source}'"))?;
+
+        let repo_dir = install_dir.join(&repo.repo_name);
+        let commit_sha = repo.commit_sha.clone();
+        let commit_url = commit_sha
+            .as_ref()
+            .and_then(|sha| commit_url_for_source(source, sha));
+        let commit_age_days = commit_age_days(local_repo_head_timestamp_ms(&repo_dir));
+
+        // Route by format: SKILL.md repos parse the file; others use format adapters.
+        match repo.format {
+            moltis_skills::formats::PluginFormat::Skill => {
+                let skill_dir = install_dir.join(&skill_state.relative_path);
+                let skill_md = skill_dir.join("SKILL.md");
+                let raw = std::fs::read_to_string(&skill_md)
+                    .map_err(|e| format!("failed to read SKILL.md: {e}"))?;
+                let content = moltis_skills::parse::parse_skill(&raw, &skill_dir)
+                    .map_err(|e| format!("failed to parse SKILL.md: {e}"))?;
+                let elig = check_requirements(&content.metadata);
+                let meta_json = moltis_skills::parse::read_meta_json(&skill_dir);
+                let display_name = meta_json.as_ref().and_then(|m| m.display_name.clone());
+                let author = meta_json.as_ref().and_then(|m| m.owner.clone());
+                let version = meta_json
+                    .as_ref()
+                    .and_then(|m| m.latest.as_ref())
+                    .and_then(|l| l.version.clone());
+                let license_url =
+                    license_url_for_source(source, content.metadata.license.as_deref());
+                let source_url: Option<String> = {
+                    let rel = &skill_state.relative_path;
+                    rel.strip_prefix(&repo.repo_name)
+                        .and_then(|p| p.strip_prefix('/'))
+                        .map(|path_in_repo| {
+                            if source.starts_with("https://") || source.starts_with("http://") {
+                                format!(
+                                    "{}/tree/main/{}",
+                                    source.trim_end_matches('/'),
+                                    path_in_repo
+                                )
+                            } else {
+                                format!("https://github.com/{}/tree/main/{}", source, path_in_repo)
+                            }
+                        })
+                };
+                Ok(serde_json::json!({
+                    "name": content.metadata.name,
+                    "display_name": display_name,
+                    "description": content.metadata.description,
+                    "author": author,
+                    "homepage": content.metadata.homepage,
+                    "version": version,
+                    "license": content.metadata.license,
+                    "license_url": license_url,
+                    "compatibility": content.metadata.compatibility,
+                    "allowed_tools": content.metadata.allowed_tools,
+                    "requires": content.metadata.requires,
+                    "eligible": elig.eligible,
+                    "missing_bins": elig.missing_bins,
+                    "install_options": elig.install_options,
+                    "trusted": skill_state.trusted,
+                    "enabled": skill_state.enabled,
+                    "quarantined": repo.quarantined,
+                    "quarantine_reason": repo.quarantine_reason,
+                    "provenance": repo.provenance,
+                    "drifted": drifted_sources.contains(source),
+                    "commit_sha": commit_sha,
+                    "commit_url": commit_url,
+                    "commit_age_days": commit_age_days,
+                    "source_url": source_url,
+                    "body": content.body,
+                    "body_html": markdown_to_html(&content.body),
+                    "source": source,
+                }))
+            },
+            format => {
+                // Non-SKILL.md format: use adapter to scan for skill body + metadata.
+                let entries = moltis_skills::formats::scan_with_adapter(&repo_dir, format)
+                    .ok_or_else(|| format!("no adapter for format '{format}'"))?
+                    .map_err(|e| format!("scan error: {e}"))?;
+                let entry = entries
+                    .into_iter()
+                    .find(|e| e.metadata.name == skill_name)
+                    .ok_or_else(|| format!("skill '{skill_name}' not found on disk"))?;
+                let source_url: Option<String> = entry.source_file.as_ref().map(|file| {
+                    if source.starts_with("https://") || source.starts_with("http://") {
+                        format!("{}/blob/main/{}", source.trim_end_matches('/'), file)
+                    } else {
+                        format!("https://github.com/{}/blob/main/{}", source, file)
+                    }
+                });
+                let license_url = license_url_for_source(source, entry.metadata.license.as_deref());
+                let empty: Vec<String> = Vec::new();
+                Ok(serde_json::json!({
+                    "name": entry.metadata.name,
+                    "display_name": entry.display_name,
+                    "description": entry.metadata.description,
+                    "author": entry.author,
+                    "homepage": entry.metadata.homepage,
+                    "version": null,
+                    "license": entry.metadata.license,
+                    "license_url": license_url,
+                    "compatibility": entry.metadata.compatibility,
+                    "allowed_tools": entry.metadata.allowed_tools,
+                    "requires": entry.metadata.requires,
+                    "eligible": true,
+                    "missing_bins": empty,
+                    "install_options": empty,
+                    "trusted": skill_state.trusted,
+                    "enabled": skill_state.enabled,
+                    "quarantined": repo.quarantined,
+                    "quarantine_reason": repo.quarantine_reason,
+                    "provenance": repo.provenance,
+                    "drifted": drifted_sources.contains(source),
+                    "commit_sha": commit_sha,
+                    "commit_url": commit_url,
+                    "commit_age_days": commit_age_days,
+                    "source_url": source_url,
+                    "body": entry.body,
+                    "body_html": markdown_to_html(&entry.body),
+                    "source": source,
+                }))
+            },
+        }
+    }
+
+    async fn install_dep(&self, params: Value) -> ServiceResult {
+        use {
+            moltis_skills::{
+                discover::{FsSkillDiscoverer, SkillDiscoverer},
+                requirements::{check_requirements, install_command_preview, run_install},
+            },
+            moltis_tools::approval::{
+                ApprovalAction, ApprovalManager, ApprovalMode, SecurityLevel,
+            },
+        };
+
+        let skill_name = params
+            .get("skill")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'skill' parameter".to_string())?;
+        let index = params.get("index").and_then(|v| v.as_u64()).unwrap_or(0) as usize;
+        let confirm = params
+            .get("confirm")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
+        let allow_host_install = params
+            .get("allow_host_install")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
+        let allow_risky_install = params
+            .get("allow_risky_install")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
+
+        // Discover the skill to get its requirements
+        let fs_discoverer = FsSkillDiscoverer::new(FsSkillDiscoverer::default_paths());
+
+        #[cfg(feature = "bundled-skills")]
+        let skills = {
+            let bundled = Arc::new(moltis_skills::bundled::BundledSkillStore::new());
+            let composite = moltis_skills::discover::CompositeSkillDiscoverer::new(
+                Box::new(fs_discoverer),
+                bundled,
+            );
+            composite.discover().await.map_err(ServiceError::message)?
+        };
+        #[cfg(not(feature = "bundled-skills"))]
+        let skills = fs_discoverer
+            .discover()
+            .await
+            .map_err(ServiceError::message)?;
+
+        let meta = skills
+            .iter()
+            .find(|s| s.name == skill_name)
+            .ok_or_else(|| format!("skill '{skill_name}' not found"))?;
+
+        let elig = check_requirements(meta);
+        let spec = elig
+            .install_options
+            .get(index)
+            .ok_or_else(|| format!("install option index {index} out of range"))?;
+
+        let command_preview = install_command_preview(spec).map_err(ServiceError::message)?;
+        if !confirm {
+            return Err(format!(
+                "dependency install requires explicit confirmation. Re-run with confirm=true after reviewing command: {command_preview}"
+            )
+            .into());
+        }
+
+        if let Some(reason) = risky_install_pattern(&command_preview)
+            && !allow_risky_install
+        {
+            security_audit(
+                "skills.install_dep_blocked",
+                serde_json::json!({
+                    "skill": skill_name,
+                    "command": command_preview,
+                    "reason": reason,
+                }),
+            );
+            return Err(format!(
+                "dependency install blocked as risky ({reason}). Re-run with allow_risky_install=true only after manual review"
+            )
+            .into());
+        }
+
+        let config = moltis_config::discover_and_load();
+        if config.tools.exec.sandbox.mode == "off" && !allow_host_install {
+            return Err("dependency install blocked because sandbox mode is off. Enable sandbox or re-run with allow_host_install=true and confirm=true".into());
+        }
+
+        let mut approval = ApprovalManager::default();
+        approval.mode =
+            ApprovalMode::parse(&config.tools.exec.approval_mode).unwrap_or(ApprovalMode::OnMiss);
+        approval.security_level = SecurityLevel::parse(&config.tools.exec.security_level)
+            .unwrap_or(SecurityLevel::Allowlist);
+        approval.allowlist = config.tools.exec.allowlist;
+
+        match approval
+            .check_command(&command_preview)
+            .await
+            .map_err(ServiceError::message)?
+        {
+            ApprovalAction::Proceed => {},
+            // skills.install_dep is an interactive RPC invoked by the user in the UI;
+            // `confirm=true` is treated as the explicit approval for this action.
+            ApprovalAction::NeedsApproval => {},
+        }
+
+        let result = run_install(spec).await.map_err(ServiceError::message)?;
+
+        security_audit(
+            "skills.install_dep",
+            serde_json::json!({
+                "skill": skill_name,
+                "command": command_preview,
+                "success": result.success,
+            }),
+        );
+
+        if result.success {
+            Ok(serde_json::json!({
+                "success": true,
+                "stdout": result.stdout,
+                "stderr": result.stderr,
+            }))
+        } else {
+            Err(format!(
+                "install failed: {}",
+                if result.stderr.is_empty() {
+                    result.stdout
+                } else {
+                    result.stderr
+                }
+            )
+            .into())
+        }
+    }
+
+    async fn skill_save(&self, params: Value) -> ServiceResult {
+        let name = params
+            .get("name")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'name' parameter".to_string())?;
+        let description = params
+            .get("description")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'description' parameter".to_string())?;
+        let body = params
+            .get("body")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'body' parameter".to_string())?;
+        let allowed_tools: Vec<String> = params
+            .get("allowed_tools")
+            .and_then(|v| v.as_array())
+            .map(|arr| {
+                arr.iter()
+                    .filter_map(|v| v.as_str().map(String::from))
+                    .collect()
+            })
+            .unwrap_or_default();
+
+        if !moltis_skills::parse::validate_name(name) {
+            return Err(format!(
+                "invalid skill name '{name}': must be 1-64 lowercase alphanumeric/hyphen chars"
+            )
+            .into());
+        }
+
+        let skills_dir = moltis_config::data_dir().join("skills");
+        let skill_dir = skills_dir.join(name);
+
+        // Build SKILL.md content.
+        let mut content = format!("---\nname: {name}\ndescription: {description}\n");
+        if !allowed_tools.is_empty() {
+            content.push_str("allowed_tools:\n");
+            for tool in &allowed_tools {
+                content.push_str(&format!("  - {tool}\n"));
+            }
+        }
+        content.push_str("---\n\n");
+        content.push_str(body);
+        if !body.ends_with('\n') {
+            content.push('\n');
+        }
+
+        std::fs::create_dir_all(&skill_dir)
+            .map_err(|e| format!("failed to create skill directory: {e}"))?;
+        std::fs::write(skill_dir.join("SKILL.md"), &content)
+            .map_err(|e| format!("failed to write SKILL.md: {e}"))?;
+
+        // Determine if this was a create or update for the response.
+        Ok(serde_json::json!({
+            "saved": true,
+            "name": name,
+            "source": "personal",
+            "path": skill_dir.to_string_lossy(),
+        }))
+    }
+
+    async fn security_status(&self) -> ServiceResult {
+        let installed_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let mcp_scan_available = command_available("mcp-scan").await;
+        let uvx_available = command_available("uvx").await;
+        Ok(serde_json::json!({
+            "mcp_scan_available": mcp_scan_available,
+            "uvx_available": uvx_available,
+            "supported": mcp_scan_available || uvx_available,
+            "installed_skills_dir": installed_dir,
+            "install_hint": "Install uv (https://docs.astral.sh/uv/) or mcp-scan to run skill security scans",
+        }))
+    }
+
+    async fn security_scan(&self) -> ServiceResult {
+        let installed_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        if !installed_dir.exists() {
+            return Ok(serde_json::json!({
+                "ok": true,
+                "message": "No installed skills directory found",
+                "results": null,
+            }));
+        }
+
+        let status = self.security_status().await?;
+        let supported = status
+            .get("supported")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
+        if !supported {
+            return Err("mcp-scan is not available. Install uvx or mcp-scan binary first".into());
+        }
+
+        let results = run_mcp_scan(&installed_dir)
+            .await
+            .map_err(ServiceError::message)?;
+        security_audit(
+            "skills.security.scan",
+            serde_json::json!({ "installed_dir": installed_dir, "status": "ok" }),
+        );
+        Ok(serde_json::json!({
+            "ok": true,
+            "installed_skills_dir": installed_dir,
+            "results": results,
+        }))
+    }
+
+    async fn recipe(&self, params: Value) -> ServiceResult {
+        let source = params
+            .get("source")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'source' parameter".to_string())?;
+        match moltis_skills::recipes::get_recipe(source) {
+            Some(recipe) => Ok(serde_json::json!({
+                "found": true,
+                "recipe": recipe,
+            })),
+            None => Ok(serde_json::json!({ "found": false })),
+        }
+    }
+
+    async fn clawhub_search(&self, params: Value) -> ServiceResult {
+        use moltis_skills::clawhub::EnrichedSearchResult;
+
+        let query = params
+            .get("query")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'query' parameter".to_string())?;
+        let client = moltis_skills::clawhub::ClawHubClient::new();
+        let response = client.search(query).await.map_err(ServiceError::message)?;
+
+        // Enrich results with stats from skill info. Use a semaphore to limit
+        // concurrent requests (avoid hitting ClawHub's 180 req/min rate limit).
+        let semaphore = Arc::new(tokio::sync::Semaphore::new(5));
+        let futs: Vec<_> = response
+            .results
+            .iter()
+            .map(|r| {
+                let slug = r.slug.clone();
+                let client = moltis_skills::clawhub::ClawHubClient::new();
+                let sem = Arc::clone(&semaphore);
+                async move {
+                    let _permit = sem.acquire().await;
+                    (slug.clone(), client.skill_info(&slug).await.ok())
+                }
+            })
+            .collect();
+        let infos = futures::future::join_all(futs).await;
+
+        let enriched: Vec<EnrichedSearchResult> = response
+            .results
+            .into_iter()
+            .map(|r| {
+                let mut e = EnrichedSearchResult::from(r.clone());
+                if let Some((_, Some(info))) = infos.iter().find(|(s, _)| *s == r.slug) {
+                    if let Some(stats) = &info.skill.stats {
+                        e.downloads = stats.downloads;
+                        e.stars = stats.stars;
+                    }
+                    if let Some(owner) = &info.owner {
+                        e.owner_handle = owner.handle.clone();
+                        e.owner_image = owner.image.clone();
+                    }
+                }
+                e
+            })
+            .collect();
+
+        Ok(serde_json::json!({ "results": enriched }))
+    }
+
+    async fn clawhub_scan(&self, params: Value) -> ServiceResult {
+        let slug = params
+            .get("slug")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'slug' parameter".to_string())?;
+        moltis_skills::clawhub::validate_slug(slug).map_err(ServiceError::message)?;
+        let client = moltis_skills::clawhub::ClawHubClient::new();
+        let scan = client.scan(slug).await.map_err(ServiceError::message)?;
+        Ok(serde_json::to_value(scan).unwrap_or_default())
+    }
+
+    async fn clawhub_info(&self, params: Value) -> ServiceResult {
+        let slug = params
+            .get("slug")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'slug' parameter".to_string())?;
+        moltis_skills::clawhub::validate_slug(slug).map_err(ServiceError::message)?;
+        let client = moltis_skills::clawhub::ClawHubClient::new();
+        let info = client
+            .skill_info(slug)
+            .await
+            .map_err(ServiceError::message)?;
+        Ok(serde_json::to_value(info).unwrap_or_default())
+    }
+
+    async fn clawhub_install(&self, params: Value) -> ServiceResult {
+        let slug = params
+            .get("slug")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| "missing 'slug' parameter".to_string())?;
+        let install_dir =
+            moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+        let skills = moltis_skills::clawhub::install_from_clawhub(slug, &install_dir)
+            .await
+            .map_err(ServiceError::message)?;
+        let installed: Vec<_> = skills
+            .iter()
+            .map(|m| {
+                serde_json::json!({
+                    "name": m.name,
+                    "description": m.description,
+                    "path": m.path.to_string_lossy(),
+                })
+            })
+            .collect();
+        security_audit(
+            "skills.clawhub.install",
+            serde_json::json!({ "slug": slug, "installed_count": installed.len() }),
+        );
+        Ok(serde_json::json!({ "installed": installed }))
+    }
+}
+
+fn local_repo_head_sha(repo_dir: &Path) -> Option<String> {
+    let repo = gix::open(repo_dir).ok()?;
+    let obj = repo.rev_parse_single("HEAD").ok()?;
+    Some(obj.detach().to_hex().to_string())
+}
+
+fn detect_and_mark_repo_drift(
+    manifest: &mut moltis_skills::types::SkillsManifest,
+    install_dir: &Path,
+) -> (bool, HashSet<String>) {
+    let mut changed = false;
+    let mut drifted = HashSet::new();
+
+    for repo in &mut manifest.repos {
+        let Some(expected_sha) = repo.commit_sha.clone() else {
+            continue;
+        };
+
+        let repo_dir = install_dir.join(&repo.repo_name);
+        let Some(current_sha) = local_repo_head_sha(&repo_dir) else {
+            continue;
+        };
+
+        if current_sha != expected_sha {
+            drifted.insert(repo.source.clone());
+            repo.commit_sha = Some(current_sha);
+            for skill in &mut repo.skills {
+                skill.trusted = false;
+                skill.enabled = false;
+            }
+            security_audit(
+                "skills.source_drift_detected",
+                serde_json::json!({
+                    "source": repo.source,
+                    "new_commit_sha": repo.commit_sha,
+                }),
+            );
+            changed = true;
+        }
+    }
+
+    (changed, drifted)
+}
+
+/// Delete a personal or project skill directory to disable it.
+fn delete_discovered_skill(source_type: &str, params: &Value) -> ServiceResult {
+    let skill_name = params
+        .get("skill")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| "missing 'skill' parameter".to_string())?;
+
+    if is_protected_discovered_skill(skill_name) {
+        return Err(ServiceError::forbidden(format!(
+            "skill '{skill_name}' is protected and cannot be deleted from the UI"
+        )));
+    }
+
+    if !moltis_skills::parse::validate_name(skill_name) {
+        return Err(format!("invalid skill name '{skill_name}'").into());
+    }
+
+    let search_dir = if source_type == "personal" {
+        moltis_config::data_dir().join("skills")
+    } else {
+        moltis_config::data_dir().join(".moltis/skills")
+    };
+
+    let skill_dir = search_dir.join(skill_name);
+    if !skill_dir.exists() {
+        return Err(format!("skill '{skill_name}' not found").into());
+    }
+
+    std::fs::remove_dir_all(&skill_dir)
+        .map_err(|e| format!("failed to delete skill '{skill_name}': {e}"))?;
+
+    security_audit(
+        "skills.discovered.delete",
+        serde_json::json!({
+            "source": source_type,
+            "skill": skill_name,
+        }),
+    );
+
+    Ok(serde_json::json!({ "source": source_type, "skill": skill_name, "deleted": true }))
+}
+
+/// Load skill detail for a personal or project skill by name.
+fn skill_detail_discovered(source_type: &str, skill_name: &str) -> ServiceResult {
+    use moltis_skills::requirements::check_requirements;
+
+    // Build search paths for the requested source type.
+    let search_dir = if source_type == "personal" {
+        moltis_config::data_dir().join("skills")
+    } else {
+        moltis_config::data_dir().join(".moltis/skills")
+    };
+
+    let skill_dir = search_dir.join(skill_name);
+    let skill_md = skill_dir.join("SKILL.md");
+    let raw = std::fs::read_to_string(&skill_md)
+        .map_err(|e| format!("failed to read SKILL.md for '{skill_name}': {e}"))?;
+
+    let content = moltis_skills::parse::parse_skill(&raw, &skill_dir)
+        .map_err(|e| format!("failed to parse SKILL.md: {e}"))?;
+
+    let elig = check_requirements(&content.metadata);
+
+    Ok(serde_json::json!({
+        "name": content.metadata.name,
+        "description": content.metadata.description,
+        "license": content.metadata.license,
+        "license_url": license_url_for_source(source_type, content.metadata.license.as_deref()),
+        "compatibility": content.metadata.compatibility,
+        "allowed_tools": content.metadata.allowed_tools,
+        "requires": content.metadata.requires,
+        "eligible": elig.eligible,
+        "missing_bins": elig.missing_bins,
+        "install_options": elig.install_options,
+        "trusted": true,
+        "enabled": true,
+        "protected": is_protected_discovered_skill(skill_name),
+        "body": content.body,
+        "body_html": markdown_to_html(&content.body),
+        "source": source_type,
+        "path": skill_dir.to_string_lossy(),
+    }))
+}
+
+/// Load skill detail for a bundled skill by name.
+#[cfg(feature = "bundled-skills")]
+fn skill_detail_bundled(skill_name: &str) -> ServiceResult {
+    use moltis_skills::requirements::check_requirements;
+
+    let store = moltis_skills::bundled::BundledSkillStore::new();
+    let skills = store.discover();
+    let meta = skills
+        .iter()
+        .find(|s| s.name == skill_name)
+        .ok_or_else(|| format!("bundled skill '{skill_name}' not found"))?;
+
+    let body = store
+        .read_skill(skill_name)
+        .ok_or_else(|| format!("bundled skill '{skill_name}' body not readable"))?;
+
+    let elig = check_requirements(meta);
+
+    Ok(serde_json::json!({
+        "name": meta.name,
+        "description": meta.description,
+        "category": meta.category,
+        "license": meta.license,
+        "compatibility": meta.compatibility,
+        "allowed_tools": meta.allowed_tools,
+        "requires": meta.requires,
+        "origin": meta.origin,
+        "eligible": elig.eligible,
+        "missing_bins": elig.missing_bins,
+        "install_options": elig.install_options,
+        "trusted": true,
+        "enabled": true,
+        "protected": true,
+        "body": body,
+        "body_html": markdown_to_html(&body),
+        "source": "bundled",
+    }))
+}
+
+fn toggle_skill(params: &Value, enabled: bool) -> ServiceResult {
+    let source = params
+        .get("source")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| "missing 'source' parameter".to_string())?;
+    let skill_name = params
+        .get("skill")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| "missing 'skill' parameter".to_string())?;
+
+    let manifest_path =
+        moltis_skills::manifest::ManifestStore::default_path().map_err(ServiceError::message)?;
+    let store = moltis_skills::manifest::ManifestStore::new(manifest_path);
+    let mut manifest = store.load().map_err(ServiceError::message)?;
+
+    let install_dir =
+        moltis_skills::install::default_install_dir().map_err(ServiceError::message)?;
+    let (drift_changed, drifted_sources) = detect_and_mark_repo_drift(&mut manifest, &install_dir);
+    if drift_changed {
+        store.save(&manifest).map_err(ServiceError::message)?;
+    }
+
+    if enabled {
+        let quarantined = manifest
+            .find_repo(source)
+            .map(|repo| repo.quarantined)
+            .ok_or_else(|| format!("repo '{source}' not found"))?;
+        if quarantined {
+            return Err(format!(
+                "repo '{source}' is quarantined. Review it and run skills.repos.unquarantine before enabling"
+            )
+            .into());
+        }
+
+        if drifted_sources.contains(source) {
+            return Err(format!(
+                "skill '{skill_name}' source changed since it was last trusted. Review and run skills.skill.trust before enabling"
+            )
+            .into());
+        }
+
+        let trusted = manifest
+            .find_repo(source)
+            .and_then(|r| r.skills.iter().find(|s| s.name == skill_name))
+            .map(|s| s.trusted)
+            .ok_or_else(|| format!("skill '{skill_name}' not found in repo '{source}'"))?;
+        if !trusted {
+            return Err(format!(
+                "skill '{skill_name}' is not trusted. Review it and run skills.skill.trust before enabling"
+            )
+            .into());
+        }
+    }
+
+    if !manifest.set_skill_enabled(source, skill_name, enabled) {
+        return Err(format!("skill '{skill_name}' not found in repo '{source}'").into());
+    }
+    store.save(&manifest).map_err(ServiceError::message)?;
+
+    security_audit(
+        "skills.skill.toggle",
+        serde_json::json!({
+            "source": source,
+            "skill": skill_name,
+            "enabled": enabled,
+        }),
+    );
+
+    Ok(serde_json::json!({ "source": source, "skill": skill_name, "enabled": enabled }))
+}
+
+fn set_skill_trusted(params: &Value, trusted: bool) -> ServiceResult {
+    let source = params
+        .get("source")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| "missing 'source' parameter".to_string())?;
+    let skill_name = params
+        .get("skill")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| "missing 'skill' parameter".to_string())?;
+
+    let manifest_path =
+        moltis_skills::manifest::ManifestStore::default_path().map_err(ServiceError::message)?;
+    let store = moltis_skills::manifest::ManifestStore::new(manifest_path);
+    let mut manifest = store.load().map_err(ServiceError::message)?;
+
+    if !manifest.set_skill_trusted(source, skill_name, trusted) {
+        return Err(format!("skill '{skill_name}' not found in repo '{source}'").into());
+    }
+
+    if !trusted {
+        let _ = manifest.set_skill_enabled(source, skill_name, false);
+    }
+
+    store.save(&manifest).map_err(ServiceError::message)?;
+    security_audit(
+        "skills.skill.trust",
+        serde_json::json!({
+            "source": source,
+            "skill": skill_name,
+            "trusted": trusted,
+        }),
+    );
+    Ok(serde_json::json!({ "source": source, "skill": skill_name, "trusted": trusted }))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn risky_install_pattern_detects_piped_shell() {
+        assert_eq!(
+            risky_install_pattern("curl https://example.com/install.sh | sh"),
+            Some("piped shell execution")
+        );
+    }
+
+    #[test]
+    fn risky_install_pattern_allows_plain_package_install() {
+        assert_eq!(risky_install_pattern("cargo install ripgrep"), None);
+    }
+}
